# Conflicts:
#	Source/RouterMobile/mobile/src/main/AndroidManifest.xml
#	Source/RouterMobile/wear/src/main/java/com/fpt/router/wear/config/DataLayerListenerService.java

+++ Source/router-service/artifacter/src/main/java/com/fpt/router/artifacter/model/algorithm/Trip.java
@@ -12,7 +12,7 @@ import java.util.List;
  */
 public class Trip implements IAlgorithmModel {
 
-    public long tripId;
+    public int tripId;
 
     public int tripNo;
